backend/requirements.txt
fastapi
uvicorn
python-multipart
httpx
python-dotenv
pydantic
gradio_client
replicate
